alot of changes again

+++ src/store/AuthContext.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
 import { createContext, useState, useEffect } from 'react';
-import { auth, db } from '../config/firebase';
+import { auth, db, messaging } from '../config/firebase';
 
 export const AuthContext = createContext<{
   user: firebase.User | null;
   setUser: React.Dispatch<React.SetStateAction<firebase.User | null>>;
-  username: string | null;
-  setUsername: React.Dispatch<React.SetStateAction<string | null>>;
-  verified: boolean;
+  username?: string | null;
+  setUsername: React.Dispatch<React.SetStateAction<string | null | undefined>>;
+  verified?: boolean;
 } | null>(null);
 
 const AuthContextProvider: React.FC = ({ children }) => {
@@ -16,10 +16,9 @@ const AuthContextProvider: React.FC = ({ children }) => {
   );
   const [user, setUser] = useState<firebase.User | null>(localUser);
 
-  const localUsername: string | null = localStorage.getItem('username');
-  const [username, setUsername] = useState<string | null>(localUsername);
-  const localVerified: boolean = !!localStorage.getItem('verified');
-  const [verified, setVerified] = useState<boolean>(localVerified);
+  const [username, setUsername] = useState<string | null>();
+
+  const [verified, setVerified] = useState<boolean>();
 
   useEffect(() => {
     const unsub = auth.onAuthStateChanged(user => {
@@ -30,8 +29,22 @@ const AuthContextProvider: React.FC = ({ children }) => {
         localStorage.removeItem('user');
         localStorage.removeItem('username');
         localStorage.removeItem('verified');
+        messaging.deleteToken();
         return;
-      };
+      }
+
+      messaging
+        .requestPermission()
+        .then(() => messaging.getToken())
+        .then(token =>
+          db.collection('devices').doc(token).set({
+            userId: auth.currentUser?.uid,
+            token
+          })
+        )
+        .catch(err => {
+          console.error(err);
+        });
 
       db.collection('usernames')
         .where('userId', '==', user.uid)
@@ -39,19 +52,16 @@ const AuthContextProvider: React.FC = ({ children }) => {
         .then(snap => {
           const username = snap.docs[0]?.id ?? null;
           setUsername(username);
-          if (username)
-            localStorage.setItem('username', username);
+          if (username) localStorage.setItem('username', username);
         });
-      
+
       db.collection('verified_users')
         .doc(user.uid)
         .get()
         .then(({ exists }) => {
-          setVerified(exists)
-          if (exists)
-            localStorage.setItem('verified', '1');
-          else
-            localStorage.removeItem('verified');
+          setVerified(exists);
+          if (exists) localStorage.setItem('verified', '1');
+          else localStorage.removeItem('verified');
         });
     });
 
@@ -59,7 +69,9 @@ const AuthContextProvider: React.FC = ({ children }) => {
   }, []);
 
   return (
-    <AuthContext.Provider value={{ user, setUser, username, setUsername, verified }}>
+    <AuthContext.Provider
+      value={{ user, setUser, username, setUsername, verified }}
+    >
       {children}
     </AuthContext.Provider>
   );
